add linting

+++ include/buildpack.bash
@@ -1,8 +1,7 @@
 
 _envfile-parse() {
     declare desc="Parse input into shell export commands"
-    local key
-    local value
+    local key value
     while read -r line || [[ -n "$line" ]]; do
         [[ "$line" =~ ^#.* ]] && continue
         [[ "$line" =~ ^$ ]] && continue
@@ -35,29 +34,39 @@ buildpack-install() {
 	declare url="$1" commit="$2" name="$3"
 	ensure-paths
 	if [[ ! "$url" ]]; then
-		asset-cat include/buildpacks.txt | while read name url commit; do
+		# shellcheck disable=SC2030
+		asset-cat include/buildpacks.txt | while read -r name url commit; do
 			buildpack-install "$url" "$commit" "$name"
 		done
 		return
 	fi
-	local target_path="$buildpack_path/${name:-$(basename $url)}"
+	# shellcheck disable=SC2031
+	# shellcheck disable=SC2154
+	local target_path="$buildpack_path/${name:-$(basename "$url")}"
+	# shellcheck disable=SC2031
 	if [[ "$(git ls-remote "$url" &> /dev/null; echo $?)" -eq 0 ]]; then
+		# shellcheck disable=SC2031
 		if [[ "$commit" ]]; then
+			# shellcheck disable=SC2031
 			if ! git clone --branch "$commit" --quiet --depth 1 "$url" "$target_path" &>/dev/null; then
 				# if the shallow clone failed partway through, clean up and try a full clone
 				rm -rf "$target_path"
+				# shellcheck disable=SC2031
 				git clone "$url" "$target_path"
-				cd "$target_path"
+				cd "$target_path" || return 1
+				# shellcheck disable=SC2031
 				git checkout --quiet "$commit"
-				cd - > /dev/null
+				cd - > /dev/null || return 1
 			else
 				echo "Cloning into '$target_path'..."
 			fi
 		else
+			# shellcheck disable=SC2031
 			git clone --depth=1 "$url" "$target_path"
 		fi
 	else
 		local tar_args
+		# shellcheck disable=SC2031
 		case "$url" in
 			*.tgz|*.tar.gz)
 				target_path="${target_path//.tgz}"
@@ -74,8 +83,10 @@ buildpack-install() {
 				tar_args="-xC"
 			;;
 		esac
+		# shellcheck disable=SC2031
 		echo "Downloading '$url' into '$target_path'..."
 		mkdir -p "$target_path"
+		# shellcheck disable=SC2031
 		curl -s --retry 2 "$url" | tar "$tar_args" "$target_path"
 		chown -R root:root "$target_path"
 		chmod 755 "$target_path"
@@ -91,14 +102,18 @@ buildpack-list() {
 
 buildpack-setup() {
 	# Buildpack expectations
+	# shellcheck disable=SC2154
 	export APP_DIR="$app_path"
 	export HOME="$app_path"
 	export REQUEST_ID="build-$RANDOM"
 	export STACK="cedar-14"
+	# shellcheck disable=SC2154
 	cp -r "$app_path/." "$build_path"
 
 	# Prepare dropped privileges
+	# shellcheck disable=SC2154
 	usermod --home "$HOME" "$unprivileged_user" > /dev/null 2>&1
+	# shellcheck disable=SC2154
 	chown -R "$unprivileged_user:$unprivileged_group" \
 		"$app_path" \
 		"$build_path" \
@@ -112,6 +127,7 @@ buildpack-setup() {
 
 	# Buildstep backwards compatibility
 	if [[ -f "$app_path/.env" ]]; then
+		# shellcheck disable=SC2046
 		eval $(cat "$app_path/.env" | _envfile-parse)
 	fi
 }
@@ -123,17 +139,17 @@ buildpack-execute() {
 		selected_path="$buildpack_path/custom"
 		rm -rf "$selected_path"
 
-		IFS='#' read url commit <<< "$BUILDPACK_URL"
+		IFS='#' read -r url commit <<< "$BUILDPACK_URL"
 		buildpack-install "$url" "$commit" custom &> /dev/null
 
 		chown -R "$unprivileged_user:$unprivileged_group" "$buildpack_path/custom"
 
-		selected_name="$(unprivileged $selected_path/bin/detect $build_path || true)"
+		selected_name="$(unprivileged "$selected_path/bin/detect" "$build_path" || true)"
 	else
 		local buildpacks=($buildpack_path/*)
 		local valid_buildpacks=()
 		for buildpack in "${buildpacks[@]}"; do
-			unprivileged $buildpack/bin/detect $build_path &> /dev/null \
+			unprivileged "$buildpack/bin/detect" "$build_path" &> /dev/null \
 				&& valid_buildpacks+=("$buildpack")
 		done
 		if [[ ${#valid_buildpacks[@]} -gt 1 ]]; then
@@ -142,7 +158,7 @@ buildpack-execute() {
 		fi
 		if [[ ${#valid_buildpacks[@]} -gt 0 ]]; then
 			selected_path="${valid_buildpacks[0]}"
-			selected_name=$(unprivileged $selected_path/bin/detect $build_path)
+			selected_name=$(unprivileged "$selected_path/bin/detect" "$build_path")
 		fi
 	fi
 	if [[ "$selected_path" ]] && [[ "$selected_name" ]]; then
@@ -152,27 +168,27 @@ buildpack-execute() {
 		exit 1
 	fi
 
-	cd "$build_path"
+	cd "$build_path" || return 1
 	unprivileged "$selected_path/bin/compile" "$build_path" "$cache_path" "$env_path"
 	if [[ -f "$selected_path/bin/release" ]]; then
 		unprivileged "$selected_path/bin/release" "$build_path" "$cache_path" > "$build_path/.release"
 	fi
 	if [[ -f "$build_path/.release" ]]; then
-		config_vars="$(cat $build_path/.release | yaml-get config_vars)"
+		config_vars="$(cat "$build_path/.release" | yaml-get config_vars)"
 		if [[ "$config_vars" ]]; then
-			mkdir -p $build_path/.profile.d
+			mkdir -p "$build_path/.profile.d"
 			OIFS=$IFS
 			IFS=$'\n'
 			for var in $config_vars; do
-				echo "export $(echo $var | sed -e 's/=/="/' -e 's/$/"/')" >> "$build_path/.profile.d/00_config_vars.sh"
+				echo "export $(echo "$var" | sed -e 's/=/="/' -e 's/$/"/')" >> "$build_path/.profile.d/00_config_vars.sh"
 			done
 			IFS=$OIFS
 		fi
 	fi
-	cd - > /dev/null
+	cd - > /dev/null || return 1
 
 	shopt -s dotglob nullglob
-	rm -rf $app_path/*
-	mv $build_path/* $app_path
+	rm -rf "${app_path:?}/*"
+	mv "${build_path:?}/*" "$app_path"
 	shopt -u dotglob nullglob
 }
